Update php-extensions.txt
This pull request proposes the inclusion of the xmlreader extension in the php-extensions.txt list for the NativePHP php-bin project. xmlreader is a lightweight, forward-only XML parser that provides an efficient way to read and process large XML documents with low memory overhead. This extension is commonly required by libraries and frameworks that rely on streaming XML data.

This change helps improve out-of-the-box support for projects using NativePHP that rely on xmlreader.